sensors: handle empty sensors
A previous commit made it so that sensors are still created when no
hooks are attached. This was needed in order to convey hook status when
every hook is ignored.

There are code paths that assume sensor/collection state exists if the
sensor was created. These locations need a way to detect if the sensor
is empty, in order to avoid logging a spurious error.

+++ pkg/sensors/load.go
@@ -85,9 +85,11 @@ func (s *Sensor) removeDirs() {
 		}
 	}
 	// Remove sensor dir
-	if err := os.Remove(filepath.Join(s.BpfDir, s.policyDir(), s.Name)); err != nil {
-		logger.GetLogger().Warn("Failed to remove sensor dir",
-			logfields.Error, err, "sensor", s.Name, "dir", filepath.Join(s.policyDir(), s.Name))
+	if !s.IsEmpty() {
+		if err := os.Remove(filepath.Join(s.BpfDir, s.policyDir(), s.Name)); err != nil {
+			logger.GetLogger().Warn("Failed to remove sensor dir",
+				logfields.Error, err, "sensor", s.Name, "dir", filepath.Join(s.policyDir(), s.Name))
+		}
 	}
 
 	// For policy dir the last one switches off the light.. there still
